Create Permutations
Write a C++ program to display all unique permutations of an input sequence of chars. For an input of
“dog”, the output should be: “dog”, “dgo”, “odg”, “ogd”, “god”, “gdo”. However, for “Bob”, the output
should be: “bob”, “bbo”, “obb”.

In this project I treated it like a state machine and noticed that in the print it moved only certain parts after x amount of turns. doing this i used a case statements that at the end of each statement a value was set to go to the next correct step.