Create MyLinkedList
Using the Node and Mylist classes made in the lecture portion. Edit the MyLists class to make the functions left in the header of the program. (a) (1 point) a function to check if the list is empty
(b) (1 point) a function to check if the list is full
(c) (2 points) an overloaded display operator
(d) (2 points) operator==
(e) (2 points) operator+
(f) (2 points) reverse. This function reverses the